tests/wpt/tests/event-timing/interactionid-orphan-pointerup.html

<!DOCTYPE html>
<html>
<meta charset=utf-8 />
<title>Event Timing: interactionId-orphan-pointerup.</title>
<button id='testButtonId'>Orphan Pointerup</button>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<script src=/resources/testdriver.js></script>
<script src=/resources/testdriver-actions.js></script>
<script src=/resources/testdriver-vendor.js></script>
<script src=resources/event-timing-test-utils.js></script>

<script>
  let observedEntries = [];
  const map = new Map();
  const events = ['pointerup'];

  promise_test(async t => {
    assert_implements(window.PerformanceEventTiming, 'Event Timing is not supported.');

    const callback = (entryList) => { observedEntries = observedEntries.concat(entryList.getEntries().filter(filterAndAddToMap(events, map))); };
    const readyToResolve = () => { return observedEntries.length >= 1; };
    const observerPromise = createPerformanceObserverPromise(['event'], callback, readyToResolve);

    await interactAndObserve('orphan-pointerup', document.getElementById('testButtonId'), observerPromise);
    assert_equals(map.get('pointerup'), 0, 'Should have a trivial interactionId for orphan pointerup event.');
  }, "Event Timing: Orphan pointerup should not be measured as an interaction.");

</script>

</html>
